+++ frontend/app/not-found.module.css
@@ -1,0 +1,20 @@
+.PageNotFound {
+  height: 60vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  font-size: clamp(1.2rem, 4vw, 2.5rem);
+  font-weight: bold;
+  text-align: center;
+
+  background: linear-gradient(
+    135deg,
+    var(--brand-aqua-teal) 0%,
+    var(--brand-warm-yellow) 100%
+  );
+  -webkit-background-clip: text;
+  background-clip: text;
+  -webkit-text-fill-color: transparent;
+  color: transparent;
+}
